Swap symlinks from root repo to ci_framework
Some jobs defined outside of the ci-framework repo call cifmw playbooks
directly (at least the adoption jobs do, not sure if there are more).
This change swamps the symlink for roles and plugins from the repo root
to the ci_framework folder, and adds others for playbooks and hooks so
those jobs will still find the playbooks and will have time to adjust.
