+++ ci_framework/playbooks
@@ -1,0 +1,1 @@
+../playbooks/
