+++ ci_framework/plugins
@@ -1,0 +1,1 @@
+../plugins/
