--- ci_framework/tests/integration/targets/make/tasks/get_makefiles_env.yml
@@ -1,31 +1,0 @@
----
-- name: Set files dir as a fact
-  ansible.builtin.set_fact:
-    get_makefiles_env_files_dir_path: >-
-      {{ ansible_env.HOME }}/.ansible/collections/ansible_collections/cifmw/general/tests/integration/targets/make/files/get_makefiles_env
-
-
-- name: "Run the target module"
-  register: get_makefiles_env_output
-  environment:
-    OUT: /home/test-user/out
-    OUTPUT_DIR: /home/test-user/output
-  cifmw.general.get_makefiles_env:
-    base_path: "{{ get_makefiles_env_files_dir_path }}/makefiles"
-
-
-- name: Debug
-  ansible.builtin.debug:
-    var: get_makefiles_env_output.makefiles_values
-
-- name: "Compare the generated values with the expected ones"
-  vars:
-    expected_variables: >-
-      {{
-        (
-          lookup('file', get_makefiles_env_files_dir_path + '/expected_variables_values.yml') |
-          from_yaml
-        ).variables
-      }}
-  ansible.builtin.assert:
-    that: expected_variables == get_makefiles_env_output.makefiles_values
