--- ci_framework/tests/unit/module_utils/test_files/networking-definition-valid-all-tools-dual-stack.yml
@@ -1,106 +1,0 @@
-networks:
-  network-1:
-    network-v4: "192.168.122.0/24"
-    network-v6: "fdc0:8b54:108a:c949::/64"
-    gateway-v4: "192.168.122.1"
-    gateway-v6: "fdc0:8b54:108a:c949::1"
-    dns-v4:
-      - "192.168.122.253"
-      - "192.168.122.254"
-    dns-v6:
-      - "fdc0:8b54:108a:c949:ffff:ffff:ffff:fffe"
-      - "fdc0:8b54:108a:c949:ffff:ffff:ffff:fffd"
-    mtu: 1500
-    tools:
-      multus:
-        ranges:
-          - start: 30
-            end: 39
-      netconfig:
-        ranges:
-          - start: "192.168.122.40"
-            end: "192.168.122.49"
-          - start: 100
-            length: 10
-      metallb:
-        ranges:
-          - start: "fdc0:8b54:108a:c949::11"
-            length: 10
-  network-3:
-    network: "fd42:add0:b7d2:09b1::/64"
-    gateway: "fd42:add0:b7d2:09b1::1"
-    dns:
-      - "fd42:add0:b7d2:09b1:ffff:ffff:ffff:fffd"
-      - "2001:4860:4860::8844"
-    vlan: 20
-    mtu: 1496
-    tools:
-      multus:
-        ranges:
-          - start: 30
-            end: 39
-      netconfig:
-        ranges:
-          - start: "fd42:add0:b7d2:09b1::64"
-            end: "fd42:add0:b7d2:09b1::f64"
-          - start: fd42:add0:b7d2:09b1::f65
-            length: 10
-      metallb:
-        ranges:
-          - start: 60
-            end: 69
-  network-2:
-    network: "172.18.0.0/24"
-    gateway-v4: "172.18.0.1"
-    dns-v4:
-      - "1.1.1.1"
-      - "172.18.0.1"
-    vlan: 21
-    mtu: 1496
-    tools:
-      netconfig:
-        ranges:
-          - start: "172.18.0.40"
-            end: "172.18.0.49"
-          - start: 100
-            length: 10
-      metallb:
-        ranges:
-          - start: 60
-            end: 69
-  network-4:
-    network-v4: "172.19.0.0/24"
-    network-v6: "fd5e:bdb2:6091:9306::/64"
-    vlan: 22
-    tools:
-      multus:
-        ranges:
-          - start: 30
-            end: 39
-      netconfig:
-        ranges:
-          - start: "172.19.0.40"
-            end: "172.19.0.49"
-          - start: 100
-            length: 10
-      metallb:
-        ranges:
-          - start: 60
-            end: 69
-group-templates:
-  group-1:
-    network-template:
-      range:
-        start: 10
-        length: 5
-    networks:
-      network-3: { }
-      network-4: { }
-      network-2: { }
-instances:
-  instance-1:
-    networks:
-      network-1:
-        ip: "192.168.122.100"
-      network-3:
-        ip: "fd42:add0:b7d2:9b1::37"
