--- ci_framework/tests/unit/module_utils/test_files/networking-definition-valid-all-tools-ipv6-only.yml
@@ -1,71 +1,0 @@
-networks:
-  network-1:
-    network-v6: "fdc0:8b54:108a:c949::/64"
-    gateway: "fdc0:8b54:108a:c949::1"
-    dns:
-      - "2001:4860:4860::8888"
-    mtu: 1500
-    tools:
-      multus:
-        ranges:
-          - start: 30
-            end: 39
-      netconfig:
-        ranges:
-          - start: 40
-            end: "90"
-          - start: 91
-            end: "fdc0:8b54:108a:c949::ff"
-      metallb:
-        ranges:
-          - start: "fdc0:8b54:108a:c949::100"
-            length: 10
-  network-3:
-    network: "fd42:add0:b7d2:09b1::/64"
-    gateway-v6: "fd42:add0:b7d2:09b1::1"
-    vlan: 20
-    mtu: 1496
-    tools:
-      multus:
-        ranges:
-          - start: 30
-            end: 99
-      netconfig:
-        ranges:
-          - start: fd42:add0:b7d2:09b1::64
-            length: 10
-      metallb:
-        ranges:
-          - start: 110
-            end: 169
-  network-2:
-    network: "fd5e:bdb2:6091:9306::/64"
-    dns-v6:
-      - "fd5e:bdb2:6091:9306:ffff:ffff:ffff:fffd"
-    vlan: 21
-    tools:
-      netconfig:
-        ranges:
-          - start: 100
-            length: 10
-      metallb:
-        ranges:
-          - start: 60
-            end: 69
-group-templates:
-  group-1:
-    network-template:
-      range:
-        start: 300
-        length: 40
-    networks:
-      network-1: { }
-      network-3: { }
-      network-2: { }
-instances:
-  instance-1:
-    networks:
-      network-1:
-        ip: "fdc0:8b54:108a:c949::f1a"
-      network-3:
-        ip: "fd42:add0:b7d2:09b1::f1a"
