--- ci_framework/tests/unit/module_utils/test_files/networking-definition-valid.yml
@@ -1,47 +1,0 @@
-networks:
-  ctlplane:
-    network: "192.168.122.0/24"
-    gateway: "192.168.122.1"
-    dns:
-      - "192.168.122.253"
-      - "192.168.122.254"
-    mtu: 1500
-  internal-api:
-    network: "172.17.0.0/24"
-    gateway: "172.17.0.1"
-    vlan: 20
-    mtu: 1496
-  storage:
-    network: "172.18.0.0/24"
-    vlan: 21
-    mtu: 1496
-  tenant:
-    network: "172.19.0.0/24"
-    gateway-v4: "172.19.0.1"
-    dns-v4:
-      - "8.8.8.8"
-      - "172.19.0.1"
-    vlan: 22
-    mtu: 1496
-group-templates:
-  crc:
-    networks:
-      ctlplane:
-        range: "192.168.122.10-192.168.122.14"
-      internal-api:
-        range: "10-14"
-      tenant:
-        range:
-          start: 10
-          length: 5
-      storage:
-        range:
-          start: 10
-          length: 5
-instances:
-  instance-1:
-    networks:
-      ctlplane:
-        ip: "192.168.122.100"
-      storage:
-        ip: "172.18.0.100"
