+++ roles/manage_secrets/molecule/dataplane_ssh/cleanup.yml
