+++ tests/unit/module_utils/test_files/networking-definition-valid-all-tools-dual-stack.yml
@@ -4,6 +4,12 @@ networks:
     network-v6: "fdc0:8b54:108a:c949::/64"
     gateway-v4: "192.168.122.1"
     gateway-v6: "fdc0:8b54:108a:c949::1"
+    dns-v4:
+      - "192.168.122.253"
+      - "192.168.122.254"
+    dns-v6:
+      - "fdc0:8b54:108a:c949:ffff:ffff:ffff:fffe"
+      - "fdc0:8b54:108a:c949:ffff:ffff:ffff:fffd"
     mtu: 1500
     tools:
       multus:
@@ -23,6 +29,9 @@ networks:
   network-3:
     network: "fd42:add0:b7d2:09b1::/64"
     gateway: "fd42:add0:b7d2:09b1::1"
+    dns:
+      - "fd42:add0:b7d2:09b1:ffff:ffff:ffff:fffd"
+      - "2001:4860:4860::8844"
     vlan: 20
     mtu: 1496
     tools:
@@ -43,6 +52,9 @@ networks:
   network-2:
     network: "172.18.0.0/24"
     gateway-v4: "172.18.0.1"
+    dns-v4:
+      - "1.1.1.1"
+      - "172.18.0.1"
     vlan: 21
     mtu: 1496
     tools:
