+++ tests/unit/module_utils/test_files/networking-definition-valid-all-tools-ipv6-only.yml
@@ -2,6 +2,8 @@ networks:
   network-1:
     network-v6: "fdc0:8b54:108a:c949::/64"
     gateway: "fdc0:8b54:108a:c949::1"
+    dns:
+      - "2001:4860:4860::8888"
     mtu: 1500
     tools:
       multus:
@@ -38,6 +40,8 @@ networks:
             end: 169
   network-2:
     network: "fd5e:bdb2:6091:9306::/64"
+    dns-v6:
+      - "fd5e:bdb2:6091:9306:ffff:ffff:ffff:fffd"
     vlan: 21
     tools:
       netconfig:
