+++ tests/unit/module_utils/test_files/networking-definition-valid.yml
@@ -2,6 +2,9 @@ networks:
   ctlplane:
     network: "192.168.122.0/24"
     gateway: "192.168.122.1"
+    dns:
+      - "192.168.122.253"
+      - "192.168.122.254"
     mtu: 1500
   internal-api:
     network: "172.17.0.0/24"
@@ -15,6 +18,9 @@ networks:
   tenant:
     network: "172.19.0.0/24"
     gateway-v4: "172.19.0.1"
+    dns-v4:
+      - "8.8.8.8"
+      - "172.19.0.1"
     vlan: 22
     mtu: 1496
 group-templates:
